Add files via upload
People study whether the cell is cancer or healthy by dozens of biochemical patterns. Chemical reactions are happening inside you while you are reading this text. Sometimes some reactions get messed, due to what unpredictable health behavior can appear. One such misbehavior is cancer. While dozens of chemical reactions change the path inside the cell due to cancer, it is a big work to track the rout of each of them. But not only chemical changes, but mechanical properties also indicate changes in cell's state. For example, in the model of mouse progeria cells are more softer than wild-type. One of the methods, we suggest, to monitor cells state and behavior is mechanical properties observation. This project has implemented the fluorescent particle-tracking technique to monitor the influence of drug compound on cancer cell.